+++ code/renderer/tr_surface.c
@@ -191,8 +191,8 @@ static void RB_SurfaceSprite( void ) {
 		float ang;
 
 		ang = M_PI * backEnd.currentEntity->e.rotation / 180;
-		s = sinf( ang );
-		c = cosf( ang );
+		s = sin( ang );
+		c = cos( ang );
 
 		VectorScale( backEnd.viewParms.or.axis[1], c * radius, left );
 		VectorMA( left, -s * radius, backEnd.viewParms.or.axis[2], left );
@@ -606,8 +606,8 @@ static void DoRailDiscs( int numSegs, const vec3_t start, const vec3_t dir, cons
 
 	for ( i = 0; i < 4; i++ )
 	{
-		c = cosf( DEG2RAD( 45 + i * 90 ) );
-		s = sinf( DEG2RAD( 45 + i * 90 ) );
+		c = cos( DEG2RAD( 45 + i * 90 ) );
+		s = sin( DEG2RAD( 45 + i * 90 ) );
 		v[0] = ( right[0] * c + up[0] * s ) * scale * spanWidth;
 		v[1] = ( right[1] * c + up[1] * s ) * scale * spanWidth;
 		v[2] = ( right[2] * c + up[2] * s ) * scale * spanWidth;
